backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "incremental-tool-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",

    # Database
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.30",
    "alembic>=1.13.1",

    # Auth / Supabase
    "python-jose[cryptography]>=3.3.0",
    "supabase>=2.4.0",

    # Statistical computation
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "scikit-learn>=1.4.0",
    "statsmodels>=0.14.2",
    "scipy>=1.13.0",

    # Async job queue
    "arq>=0.25.0",
    "redis>=5.0.0",

    # LLM (OpenRouter — OpenAI-compatible endpoint, model hot-swappable via settings)
    "openai>=1.30.0",

    # PDF generation
    "weasyprint>=62.0",
    "jinja2>=3.1.4",

    # Utilities
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.6",
    "pytest-cov>=5.0.0",
    "factory-boy>=3.3.0",
    "faker>=25.0.0",
    "respx>=0.21.0",
    "ruff>=0.4.4",
    "mypy>=1.10.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=app --cov-report=term-missing --cov-report=html"

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
